requirements.txt
alembic==1.7.6
aniso8601==9.0.1
click==8.0.3
ecdsa==0.17.0
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
python-jose==3.3.0
pytz==2021.3
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
